api/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-stubs==1.7.0
djangorestframework==3.12.1
djangorestframework-camel-case==1.2.0
idna==2.10
mypy==0.800
mypy-extensions==0.4.3
pip-autoremove==0.9.1
pysqlite3-binary==0.4.3
python-decouple==3.4
python-jsonrpc-server==0.4.0
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
testfixtures==6.15.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.25.11
git+git://github.com/matangover/mypyls@master#egg=mypyls
